Address warnings reported by clang++ in Milena.
	* mln/canvas/browsing/diagonal2d.hh (diagonal2d_t::operator())
	* mln/clustering/kmean_rgb.hh (internal::kmean_rgb_dispatch)
	* mln/labeling/value_and_compute.hh
	(internal::value_and_compute_tests)
	(impl::process__)
	* mln/metal/math/root.hh (impl::root<n, x, lo, hi>::value): Add an
	explicit type conversion of operator<'s the right-hand side
	operand.
	* mln/morpho/watershed/topological.hh (topological)
	* mln/world/rgb/invert.hh (internal::invert_dispatch):
	Remove unused variables.
	Do not name unused arguments.
	* mln/util/object_id.hh
	(operator==(const object_id<Tag,V>&, const Value<V2>&)):
	Add an explicit type conversion of the right-hand side operand.
	* mln/world/rgb/invert.hh,
	* tests/core/alias/point1d.cc,
	* tests/make/p_edges_with_mass_centers.cc,
	* tests/make/p_vertices_with_mass_centers.cc,
	* tests/value/int_s16.cc,
	* tests/value/int_u8.cc:
	Shush the compiler about unused variables.
	* tests/trait/op/plus.cc (mln::my_image2d<T>): Complete this type
	to have it comply with the mln::Image concept.
	(mln::dummy_t): New (dummy) type.
	(dummy): New (dummy) value.
	(main): Do not use uninitialized pointers.

+++ milena/mln/metal/math/root.hh
@@ -1,4 +1,5 @@
-// Copyright (C) 2007, 2008, 2009 EPITA Research and Development Laboratory (LRDE)
+// Copyright (C) 2007, 2008, 2009, 2013 EPITA Research and Development
+// Laboratory (LRDE)
 //
 // This file is part of Olena.
 //
@@ -60,7 +61,7 @@ namespace mln
 	    val_lo = root<n, x, lo, mid-1>::value,
 	    val_hi = root<n, x, mid, hi>::value
 	  };
-	  enum { value = x < mlc_pow_int(mid, n) ? val_lo : val_hi };
+	  enum { value = x < unsigned(mlc_pow_int(mid, n)) ? val_lo : val_hi };
 	};
 
 	template<unsigned n, unsigned x, unsigned m>
